refactor(console): reorg role transfer component file structure
