+++ packages/console/src/components/RolesTransfer/SourceRolesBox/index.tsx
@@ -17,8 +17,7 @@ import useDebounce from '@/hooks/use-debounce';
 import * as transferLayout from '@/scss/transfer.module.scss';
 import { buildUrl } from '@/utils/url';
 
-import SourceRoleItem from '../SourceRoleItem';
-
+import SourceRoleItem from './SourceRoleItem';
 import * as styles from './index.module.scss';
 
 type Props = {
