+++ packages/console/src/components/RolesTransfer/TargetRolesBox/index.tsx
@@ -3,8 +3,7 @@ import { useTranslation } from 'react-i18next';
 
 import * as transferLayout from '@/scss/transfer.module.scss';
 
-import TargetRoleItem from '../TargetRoleItem';
-
+import TargetRoleItem from './TargetRoleItem';
 import * as styles from './index.module.scss';
 
 type Props = {
